Fix a bug where active Tiles wouldn't be returned
